docs: update icons
