+++ projects/demo/src/app/app.component.ts
@@ -1,7 +1,7 @@
-import { Component, inject } from '@angular/core';
-import { MatIconRegistry } from '@angular/material/icon';
-import { DomSanitizer } from '@angular/platform-browser';
+import { Component } from '@angular/core';
 import { RouterOutlet } from '@angular/router';
+
+import { injectIcons } from './utils/inject-icons';
 
 @Component({
   selector: 'app-root',
@@ -12,24 +12,6 @@ import { RouterOutlet } from '@angular/router';
 })
 export class AppComponent {
   constructor() {
-    const sanitizer = inject(DomSanitizer);
-    const iconRegistry = inject(MatIconRegistry);
-
-    iconRegistry.addSvgIcon(
-      'logo',
-      sanitizer.bypassSecurityTrustResourceUrl(`/assets/angular.svg`),
-    );
-    iconRegistry.addSvgIcon(
-      'theme',
-      sanitizer.bypassSecurityTrustResourceUrl(`/assets/theme.svg`),
-    );
-    iconRegistry.addSvgIcon(
-      'github-circle',
-      sanitizer.bypassSecurityTrustResourceUrl(`/assets/github-circle.svg`),
-    );
-    iconRegistry.addSvgIcon(
-      'github-round',
-      sanitizer.bypassSecurityTrustResourceUrl(`/assets/github-round.svg`),
-    );
+    injectIcons();
   }
 }
